// src/types/schemas-auth.d.ts
export interface components {
    schemas: {
// ...
        DiscoverLoginContextsResponse: {
            selectionToken?: string;
            /** Format: int64 */
            expiresInSeconds?: number;
            contexts?: components["schemas"]["DiscoveredLoginContextResponse"][];
        };
// ...
        UserMembershipResponse: {
            /** Format: uuid */
            tenantId?: string;
            /** Format: uuid */
            organizationId?: string;
            organizationCode?: string;
            organizationName?: string;
            roleCode?: string;
            roleName?: string;
            scopeType?: string;
            permissions?: string[];
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}